Use PRC password_obfuscation code
The consistency checker included its own version of the iRODS
password obfuscation code. This was needed because this code wasn't
included in Python-irodsclient (PRC) in the past. Since it's been
included now for quite some time, we can just switch to the PRC
version and remove the local version, as well as its dependencies.